requirements.txt
google-generativeai>=0.7.0
opencv-python>=4.5.0
numpy>=1.20.0
pillow>=8.0.0
zenoh>=0.7.0 